% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newapp.R
\name{draw_court}
\alias{draw_court}
\title{Uses the play by play data to draw a NBA court with coordinates of shot locations on it}
\usage{
draw_court(dat)
}
\arguments{
\item{data}{play by play data of a NBA game}
}
\description{
Uses the play by play data to draw a NBA court with coordinates of shot locations on it
}
