% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newapp.R
\name{game_coordinates}
\alias{game_coordinates}
\title{Computes the x and y coordinates of shots, based on NBA play by play data}
\usage{
game_coordinates(data, period, time, team, player)
}
\arguments{
\item{data}{gamedata() with complete NBA play by play data}

\item{period}{input variable to dplyr::filter coordinates by period}

\item{time}{input variable to dplyr::filter coordinates by time left in period}

\item{team}{input variable to dplyr::filter coordinates by team}
}
\value{
gamedata() with adjusted x and y coordinates to fit on basketball court plot
}
\description{
Computes the x and y coordinates of shots, based on NBA play by play data
}
