% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newapp.R
\name{get_data}
\alias{get_data}
\title{Uses the game id to return the full play by play data for this id from nba-stats.com}
\usage{
get_data(id)
}
\arguments{
\item{id}{numeric, valid NBA game id}
}
\value{
df with play by play data
}
\description{
Uses the game id to return the full play by play data for this id from nba-stats.com
}
