% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newapp.R
\name{get_id}
\alias{get_id}
\title{Uses the matchup and date of a game to determine the game id}
\usage{
get_id(matchup, in_date)
}
\arguments{
\item{matchup}{based on NBA game log data, in the format "XXX ad YYY" using team tri-codes.}
}
\value{
id of the game
}
\description{
Uses the matchup and date of a game to determine the game id
}
