% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newapp.R
\name{get_logo_away}
\alias{get_logo_away}
\title{Uses the game id and matchup to provid a link to the away teams logo}
\usage{
get_logo_away(data, game_id)
}
\arguments{
\item{data}{game logs of the NBA season}

\item{game_id}{numeric, valid NBA game id}
}
\value{
src, source of link to picture of the away teams logo
}
\description{
Uses the game id and matchup to provid a link to the away teams logo
}
