% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newapp.R
\name{get_player_picture}
\alias{get_player_picture}
\title{Uses the game id and play by play data to provid a link to the players photo}
\usage{
get_player_picture(data, player)
}
\arguments{
\item{data}{play by play data of a NBA game}

\item{game_id}{numeric, valid NBA game id}
}
\value{
src, source of link to picture of the the players photo
}
\description{
Uses the game id and play by play data to provid a link to the players photo
}
