% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newapp.R
\docType{data}
\name{logs}
\alias{logs}
\title{logs Dataset Description}
\format{
A data frame with X rows and Y columns.
}
\source{
Provide information about the source of the dataset.
}
\usage{
logs
}
\description{
This is a brief description of the dataset.
}
\keyword{datasets}
