% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newapp.R
\docType{data}
\name{logs}
\alias{logs}
\title{logs Dataset Description}
\format{
A data frame
}
\source{
From the NBA stats website
}
\usage{
logs
}
\description{
The gamelogs for the 2022-23 NBA season
}
\keyword{datasets}
